+++ hydromt/data_adapter/geodataframe.py
@@ -1,5 +1,6 @@
 """The Geodataframe adapter implementation."""
 import logging
+import warnings
 from os.path import join
 from pathlib import Path
 from typing import NewType, Union
@@ -45,6 +46,7 @@ class GeoDataFrameAdapter(DataAdapter):
         driver_kwargs: dict = {},
         name: str = "",  # optional for now
         catalog_name: str = "",  # optional for now
+        **kwargs,
     ):
         """Initiate data adapter for geospatial vector data.
 
@@ -93,6 +95,14 @@ class GeoDataFrameAdapter(DataAdapter):
         name, catalog_name: str, optional
             Name of the dataset and catalog, optional for now.
         """
+        if kwargs:
+            warnings.warn(
+                "Passing additional keyword arguments to be used by the "
+                "GeoDataFrameAdapter driver is deprecated and will be removed "
+                "in a future version. Please use 'driver_kwargs' instead.",
+                DeprecationWarning,
+            )
+            driver_kwargs.update(kwargs)
         super().__init__(
             path=path,
             driver=driver,