+++ tests/test_data_catalog.py
@@ -288,3 +288,29 @@ def test_get_data(df):
     assert isinstance(df, pd.DataFrame)
     with pytest.raises(ValueError, match='Unknown tabular data type "list"'):
         data_catalog.get_dataframe([])
+
+
+def test_deprecation_warnings(artifact_data):
+    with pytest.deprecated_call():
+        # should be DataCatalog(data_libs=['artifact_data=v0.0.6'])
+        DataCatalog(artifact_data="v0.0.6")
+    with pytest.deprecated_call():
+        cat = DataCatalog()
+        # should be cat.from_predefined_catalogs('artifact_data', 'v0.0.6')
+        cat.from_artifacts("artifact_data", version="v0.0.6")
+    with pytest.deprecated_call():
+        fn = artifact_data["chelsa"].path
+        # should be driver_kwargs=dict(chunks={'x': 100, 'y': 100})
+        artifact_data.get_rasterdataset(fn, chunks={"x": 100, "y": 100})
+    with pytest.deprecated_call():
+        fn = artifact_data["gadm_level1"].path
+        # should be driver_kwargs=dict(assert_gtype='Polygon')
+        artifact_data.get_geodataframe(fn, assert_gtype="MultiPolygon")
+    with pytest.deprecated_call():
+        fn = artifact_data["grdc"].path
+        # should be driver_kwargs=dict(index_col=0)
+        artifact_data.get_dataframe(fn, index_col=0)
+    with pytest.deprecated_call():
+        fn = artifact_data["gtsmv3_eu_era5"].path
+        # should be driver_kwargs=dict(chunks={'time': 100})
+        artifact_data.get_geodataset(fn, chunks={"time": 100})